+++ app/api/models.py
@@ -2,10 +2,11 @@
 
 
 import datetime
+import jwt
 
 from flask import current_app
 
-from app import db, bcrypt
+from app.extensions import db, bcrypt
 
 
 # User model
@@ -33,3 +34,36 @@ class User(db.Model):
             password, current_app.config.get('BCRYPT_LOG_ROUNDS')
         ).decode()
         self.created_at = created_at
+
+    @classmethod
+    def encode_auth_token(cls, user_id):
+        """Generate the auth token."""
+        try:
+            days = current_app.config.get('TOKEN_EXPIRATION_DAYS')
+            seconds = current_app.config.get('TOKEN_EXPIRATION_SECONDS')
+            payload = {
+                'exp': datetime.datetime.utcnow() + datetime.timedelta(
+                    days=days, seconds=seconds),
+                'iat': datetime.datetime.utcnow(),
+                'sub': user_id
+            }
+            return jwt.encode(
+                payload,
+                current_app.config.get('SECRET_KEY'),
+                algorithm='HS256'
+            )
+        # pylint: disable=broad-except
+        except Exception as e:
+            return e
+
+    @staticmethod
+    def decode_auth_token(auth_token):
+        """Decode the auth token - :param auth_token: - :return: integer|string."""
+        try:
+            secret = current_app.config.get('SECRET_KEY')
+            payload = jwt.decode(auth_token, secret)
+            return payload['sub']
+        except jwt.ExpiredSignatureError:
+            return 'Signature expired. Please log in again.'
+        except jwt.InvalidTokenError:
+            return 'Invalid token. Please log in again.'