+++ app/config.py
@@ -12,15 +12,16 @@ class Config(object):
     TESTING = False
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     SECRET_KEY = os.environ.get('SECRET_KEY')
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
     BCRYPT_LOG_ROUNDS = 13
+    TOKEN_EXPIRATION_DAYS = 30
+    TOKEN_EXPIRATION_SECONDS = 0
 
 
 class DevelopmentConfig(Config):
     """Configurations for Development."""
 
     DEBUG = True
-    SECRET_KEY = 'my_precious'
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
     BCRYPT_LOG_ROUNDS = 4
 
 
@@ -29,21 +30,22 @@ class TestingConfig(Config):
 
     DEBUG = True
     TESTING = True
-    SECRET_KEY = 'my_precious'
     SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_TEST_URL')
     BCRYPT_LOG_ROUNDS = 4
+    TOKEN_EXPIRATION_DAYS = 0
+    TOKEN_EXPIRATION_SECONDS = 3
 
 
 class StagingConfig(Config):
     """Configurations for Staging."""
 
     DEBUG = True
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
 
 
 class ProductionConfig(Config):
     """Configurations for Production."""
 
+    # Set these explicitly just to be extra safe
     DEBUG = False
     TESTING = False
     SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')