+++ app/tests/test_config.py
@@ -22,14 +22,19 @@ class TestDevelopmentConfig(TestCase):
 
     def test_app_is_development(self):
         """Ensure appropriate configuration values for development environment."""
-        self.assertTrue(app.config['SECRET_KEY'] == 'my_precious')
         self.assertTrue(app.config['DEBUG'] is True)
         self.assertFalse(current_app is None)
+        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_DAYS'] == 30)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_SECONDS'] == 0)
+        self.assertTrue(
+            app.config['SECRET_KEY'] ==
+            os.environ.get('SECRET_KEY')
+        )
         self.assertTrue(
             app.config['SQLALCHEMY_DATABASE_URI'] ==
             os.environ.get('DATABASE_URL')
         )
-        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
 
 
 class TestTestingConfig(TestCase):
@@ -41,15 +46,20 @@ class TestTestingConfig(TestCase):
 
     def test_app_is_testing(self):
         """Ensure appropriate configuration values for testing environment."""
-        self.assertTrue(app.config['SECRET_KEY'] == 'my_precious')
         self.assertTrue(app.config['DEBUG'])
         self.assertTrue(app.config['TESTING'])
         self.assertFalse(app.config['PRESERVE_CONTEXT_ON_EXCEPTION'])
+        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_DAYS'] == 0)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_SECONDS'] == 3)
+        self.assertTrue(
+            app.config['SECRET_KEY'] ==
+            os.environ.get('SECRET_KEY')
+        )
         self.assertTrue(
             app.config['SQLALCHEMY_DATABASE_URI'] ==
             os.environ.get('DATABASE_TEST_URL')
         )
-        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
 
 
 class TestProductionConfig(TestCase):
@@ -64,6 +74,12 @@ class TestProductionConfig(TestCase):
         self.assertFalse(app.config['DEBUG'])
         self.assertFalse(app.config['TESTING'])
         self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 13)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_DAYS'] == 30)
+        self.assertTrue(app.config['TOKEN_EXPIRATION_SECONDS'] == 0)
+        self.assertTrue(
+            app.config['SECRET_KEY'] ==
+            os.environ.get('SECRET_KEY')
+        )
 
 
 if __name__ == '__main__':