+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Production release process and instructions.
 - Code coverage for testing.
 - Password to users.
+- Authentication endpoints and tests.
 
 ### Changed
 - Moved API to blueprint.